Fix GetAverageScoreOfStudySessionsPerMonthPerStack to actually get average instead of sum and able to show double type numbers

+++ SpirosZoupas.Console.Flashcards/DAL/DTO/StudySessionPerMonthDTO.cs
@@ -3,17 +3,17 @@
     public class StudySessionPerMonthDTO
     {
         public string StackName { get; set; }
-        public int January { get; set; }
-        public int February { get; set; }
-        public int March { get; set; }
-        public int April { get; set; }
-        public int May { get; set; }
-        public int June { get; set; }
-        public int July { get; set; }
-        public int August { get; set; }
-        public int September { get; set; }
-        public int October { get; set; }
-        public int November { get; set; }
-        public int December { get; set; }
+        public double January { get; set; }
+        public double February { get; set; }
+        public double March { get; set; }
+        public double April { get; set; }
+        public double May { get; set; }
+        public double June { get; set; }
+        public double July { get; set; }
+        public double August { get; set; }
+        public double September { get; set; }
+        public double October { get; set; }
+        public double November { get; set; }
+        public double December { get; set; }
     }
 }
